[Functions] Wrap shared fucntions in Class
